Extract query tests to break `query` -> `db` cycle
The tests in `src/query/hybrid.rs` and `src/query/semantic_pathfinding.rs`
were effectively integration tests that imported `crate::db::AletheiaDB`.
Since `db` itself imports from `query` (e.g. `QueryPlanner`, `QueryExecutor`),
this created a circular dependency.

This commit extracts these tests into the root `tests/` directory as
`query_hybrid_tests.rs` and `query_semantic_pathfinding_tests.rs`.
This correctly positions them as external integration tests depending
on the public API of the crate, fully severing the structural cycle.
